Abstracted Weightscheme
